Removed Dependencies.
# Version 2.7.1
I removed several dependencies (Midi-QoL, Active Auras, and Item Piles). Items in the module have not been updated to reflect the changes here yet.

## Midi-QoL
Midi-QoL is not 5E 4.0 ready and isn’t likely to be for months. Its purpose as a dependency was to automate some reminders and provide GM roll automation. However, some of these functions have been absorbed by the base D&D implementation, and the others can be fulfilled by other modules (See Remaining Dependencies).

## Active Auras
Active Auras was used by 1 paladin aura, but it can be handled as a reminder with Build-a-Bonus.

## Item Piles
This module hasn’t been seeing as much use, and is struggling to update in a timely fashion with Foundry.

## Remaining Dependencies
Going forward, the only two dependencies left in the module are for Build-a-Bonus and Effect Macro. Both modules are 5E 4.0 ready and can fulfill the majority of functions needed by this compendium module.